Modularize datamodels in Execution API
Split the `datamodels` module to granular mods
