+++ airflow/api_fastapi/execution_api/datamodels/__init__.py
@@ -14,19 +14,3 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-
-from __future__ import annotations
-
-from typing import Annotated
-
-from fastapi import Depends
-
-from airflow.api_fastapi.execution_api.datamodels.token import TIToken
-
-
-def get_task_token() -> TIToken:
-    """TODO: Placeholder for task identity authentication. This should be replaced with actual JWT decoding and validation."""
-    return TIToken(ti_key="test_key")
-
-
-TokenDep = Annotated[TIToken, Depends(get_task_token)]
