+++ airflow/api_fastapi/execution_api/datamodels/taskinstance.py
@@ -17,16 +17,9 @@
 
 from __future__ import annotations
 
-from typing import Annotated, Any, Literal, Union
+from typing import Annotated, Literal, Union
 
-from pydantic import (
-    BaseModel,
-    ConfigDict,
-    Discriminator,
-    Field,
-    Tag,
-    WithJsonSchema,
-)
+from pydantic import BaseModel, ConfigDict, Discriminator, Tag, WithJsonSchema
 
 from airflow.api_fastapi.common.types import UtcDateTime
 from airflow.utils.state import IntermediateTIState, TaskInstanceState as TIState, TerminalTIState
@@ -104,40 +97,3 @@ class TIHeartbeatInfo(BaseModel):
 
     hostname: str
     pid: int
-
-
-class ConnectionResponse(BaseModel):
-    """Connection schema for responses with fields that are needed for Runtime."""
-
-    conn_id: str
-    conn_type: str
-    host: str | None
-    schema_: str | None = Field(alias="schema")
-    login: str | None
-    password: str | None
-    port: int | None
-    extra: str | None
-
-
-class VariableResponse(BaseModel):
-    """Variable schema for responses with fields that are needed for Runtime."""
-
-    model_config = ConfigDict(from_attributes=True)
-
-    key: str
-    val: str | None = Field(alias="value")
-
-
-class XComResponse(BaseModel):
-    """XCom schema for responses with fields that are needed for Runtime."""
-
-    key: str
-    value: Any
-    """The returned XCom value in a JSON-compatible format."""
-
-
-# TODO: This is a placeholder for Task Identity Token schema.
-class TIToken(BaseModel):
-    """Task Identity Token."""
-
-    ti_key: str
