+++ airflow/api_fastapi/execution_api/datamodels/token.py
@@ -17,16 +17,11 @@
 
 from __future__ import annotations
 
-from typing import Annotated
-
-from fastapi import Depends
-
-from airflow.api_fastapi.execution_api.datamodels.token import TIToken
+from pydantic import BaseModel
 
 
-def get_task_token() -> TIToken:
-    """TODO: Placeholder for task identity authentication. This should be replaced with actual JWT decoding and validation."""
-    return TIToken(ti_key="test_key")
+# TODO: This is a placeholder for Task Identity Token schema.
+class TIToken(BaseModel):
+    """Task Identity Token."""
 
-
-TokenDep = Annotated[TIToken, Depends(get_task_token)]
+    ti_key: str
