+++ airflow/api_fastapi/execution_api/routes/task_instances.py
@@ -29,7 +29,12 @@ from sqlalchemy.sql import select
 
 from airflow.api_fastapi.common.db.common import get_session
 from airflow.api_fastapi.common.router import AirflowRouter
-from airflow.api_fastapi.execution_api import datamodels
+from airflow.api_fastapi.execution_api.datamodels.taskinstance import (
+    TIEnterRunningPayload,
+    TIHeartbeatInfo,
+    TIStateUpdate,
+    TITerminalStatePayload,
+)
 from airflow.models.taskinstance import TaskInstance as TI
 from airflow.utils import timezone
 from airflow.utils.state import State
@@ -55,7 +60,7 @@ log = logging.getLogger(__name__)
 )
 def ti_update_state(
     task_instance_id: UUID,
-    ti_patch_payload: Annotated[datamodels.TIStateUpdate, Body()],
+    ti_patch_payload: Annotated[TIStateUpdate, Body()],
     session: Annotated[Session, Depends(get_session)],
 ):
     """
@@ -85,7 +90,7 @@ def ti_update_state(
 
     query = update(TI).where(TI.id == ti_id_str).values(data)
 
-    if isinstance(ti_patch_payload, datamodels.TIEnterRunningPayload):
+    if isinstance(ti_patch_payload, TIEnterRunningPayload):
         if previous_state != State.QUEUED:
             log.warning(
                 "Can not start Task Instance ('%s') in invalid state: %s",
@@ -115,7 +120,7 @@ def ti_update_state(
             pid=ti_patch_payload.pid,
             state=State.RUNNING,
         )
-    elif isinstance(ti_patch_payload, datamodels.TITerminalStatePayload):
+    elif isinstance(ti_patch_payload, TITerminalStatePayload):
         query = TI.duration_expression_update(ti_patch_payload.end_date, query, session.bind)
 
     # TODO: Replace this with FastAPI's Custom Exception handling:
@@ -143,7 +148,7 @@ def ti_update_state(
 )
 def ti_heartbeat(
     task_instance_id: UUID,
-    ti_payload: datamodels.TIHeartbeatInfo,
+    ti_payload: TIHeartbeatInfo,
     session: Annotated[Session, Depends(get_session)],
 ):
     """Update the heartbeat of a TaskInstance to mark it as alive & still running."""
